gateway module

+++ backend/api-manager/src/main/java/com/apimanager/portal/dto/RateLimitRequest.java
@@ -4,8 +4,8 @@ import lombok.Data;
 
 @Data
 public class RateLimitRequest {
-    private Integer rateLimitPerMinute;  // null = disabled
-    private Integer rateLimitPerHour;
-    private Integer rateLimitPerDay;
-    private Integer rateLimitTotal;
+    private Long rateLimitPerMinute;  // null = disabled
+    private Long rateLimitPerHour;
+    private Long rateLimitPerDay;
+    private Long rateLimitTotal;
 }
