+++ backend/api-manager/src/main/java/com/apimanager/registry/dto/ApiResponse.java
@@ -24,10 +24,10 @@ public class ApiResponse {
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
     private List<ApiEndpointResponse> endpoints;
-    private Integer rateLimitPerMinute;
-    private Integer rateLimitPerHour;
-    private Integer rateLimitPerDay;
-    private Integer rateLimitTotal;
+    private Long rateLimitPerMinute;
+    private Long rateLimitPerHour;
+    private Long rateLimitPerDay;
+    private Long rateLimitTotal;
  
     
 }
